EDITED training skript
- in ffnn_model added following methods:
-> prediction : predict output of model
-> step : update weight
-> zero_grad : resets gradients to zero for every batch

change db in FNN to np.mean(db, axis = 1) due to dim problems ( dim is 1000,128 , but should be 128,1)

The MNIST data set has to be downloded in dir /data + the packege has to be installed

Notes:
I also did not implemented a regularization, but might be helpful.
I don't think we need to perform cross validation, since we have no hyperparameter tuning. What do you think?